Reorganized severity table to more accurately reflect model.
We noticed a discrepancy between population average CFR and the output
of our simulation.
This is due to the fact that all recovered/fatal cases must flow to the
hospital state.
Thus our simulation CFR was the product of two averages which isn't
equal to the average of the product of rates.
This restructuring allows us to more accurately compare while not having
to simulate each age class independently.