added URLSerializer util
Quick and dirty helper to serialize/deserialize parameters within the URL,
so people can share/save it and keep their parameters

This could have been done inside a redux middleware, but with some refacto.

We use JSON.stringify to serialize things out. It's not the most optimized way at all,
but it works, and it's simple enough. Note that Date object is serialized as a string,
so some extra work is needed during deserialization.