feat: link to previous scenarios (#53)
* added URLSerializer util

Quick and dirty helper to serialize/deserialize parameters within the URL,
so people can share/save it and keep their parameters

This could have been done inside a redux middleware, but with some refacto.

We use JSON.stringify to serialize things out. It's not the most optimized way at all,
but it works, and it's simple enough. Note that Date object is serialized as a string,
so some extra work is needed during deserialization.

* Main.tsx uses URLSerializer

We serialize the parameters when calling `handleSubmit`

And deserialize it using the "init" parameter of the useReducer hook
function.

If nothing is passed in the URL, then the default initScenario is set.

If something bad is passed in the URL, same (with a catched error)