validation of numberPoints in reducer
formik Field updates field value even when outside specified
min-max range (e.g. when updating with keyboard, not with arrows).
This allows for null value of numberPoints, or for value outside
desired 5-100 range, which may lead to 'Containment cannot be
empty' error in algorithms/run.ts. Added aditional validation
of numberPoints value in the reducer.