Refactor how labels/text are passed to FormHelpButton.
This removes the react-children-utilities dependency,
which fixes the tests without requiring module transpilation.

The FormHelpButton uses the label field to label the help box, so if the label is a React Node then the text is difficult to access programmatically. The previous implementation used a 3rd party dependency, react-children-utilities, to extract the text. But the dependency breaks the unit tests because the module format is not supported by Jest. This commit creates a HelpProps interface to communicate help label and text to the FormHelpButton. This is more explicit, and avoids the need for another dependency.