fix: webpack error: "module 'fs' not found"
This replaces `jsonp` package dependency of `react-share` with a dummy empty module.

`jsonp` which itself haven't been updated for years now, triggered a build error due to the outdated dependency on `debug@2` package, which happened to use `fs` back in the day, which is not available in browsers.

We don't need the `react-share` functionality associated with `jsonp` (share counts) and we want to avoid including dependencies that haven't been updated for years.

Ideally it should be fixed upstream in the `react-share`.